+++ packages/perplexity/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/perplexity
 
+## 4.0.0-beta.14
+
+### Patch Changes
+
+- c29a26f: feat(provider): add support for provider references and uploading files as supported per provider
+- Updated dependencies [c29a26f]
+  - @ai-sdk/provider-utils@5.0.0-beta.10
+  - @ai-sdk/provider@4.0.0-beta.6
+
+## 4.0.0-beta.13
+
+### Patch Changes
+
+- 38fc777: Add AI Gateway hint to provider READMEs
+
 ## 4.0.0-beta.12
 
 ### Patch Changes