+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/provider-utils
 
+## 5.0.0-beta.10
+
+### Patch Changes
+
+- c29a26f: feat(provider): add support for provider references and uploading files as supported per provider
+- Updated dependencies [c29a26f]
+  - @ai-sdk/provider@4.0.0-beta.6
+
 ## 5.0.0-beta.9
 
 ### Patch Changes