+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,32 @@
 
 ### Patch Changes
 
+- ai@7.0.0-beta.62
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [c29a26f]
+  - ai@7.0.0-beta.61
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [38fc777]
+  - ai@7.0.0-beta.60
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@7.0.0-beta.59
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [2e17091]
   - ai@7.0.0-beta.58