+++ packages/vercel/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/vercel
 
+## 3.0.0-beta.14
+
+### Patch Changes
+
+- Updated dependencies [c29a26f]
+  - @ai-sdk/openai-compatible@3.0.0-beta.14
+  - @ai-sdk/provider-utils@5.0.0-beta.10
+  - @ai-sdk/provider@4.0.0-beta.6
+
+## 3.0.0-beta.13
+
+### Patch Changes
+
+- Updated dependencies [38fc777]
+  - @ai-sdk/openai-compatible@3.0.0-beta.13
+
 ## 3.0.0-beta.12
 
 ### Patch Changes